requirements.txt
nose==1.3.7
psycopg2==2.6.1
testing.postgresql==1.2.1
